# Source/backend/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Документ API (PHP)",
    "version": "1.0.0",
    "description": "REST API реестра учётных карточек"
  },
  "servers": [{ "url": "http://localhost/kursovaya/Source/backend" }],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "paths": {
    "/auth/login": {
      "post": {
        "summary": "Авторизация",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": { "type": "string" },
                  "password": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": { "200": { "description": "JWT токен" } }
      }
    },
    "/users": {
      "get": {
        "summary": "Список записей реестра",
        "security": [{ "bearerAuth": [] }],
        "responses": { "200": { "description": "OK" } }
      },
      "post": {
        "summary": "Создать запись (admin)",
        "security": [{ "bearerAuth": [] }],
        "responses": { "201": { "description": "Created" } }
      }
    },
    "/users/{id}": {
      "get": {
        "summary": "Одна запись",
        "security": [{ "bearerAuth": [] }],
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }],
        "responses": { "200": { "description": "OK" } }
      },
      "delete": {
        "summary": "Удалить запись (admin)",
        "security": [{ "bearerAuth": [] }],
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }],
        "responses": { "200": { "description": "Deleted" } }
      }
    },
    "/reports/export": {
      "get": {
        "summary": "Экспорт отчёта JSON",
        "security": [{ "bearerAuth": [] }],
        "responses": { "200": { "description": "Report" } }
      }
    },
    "/app-users": {
      "get": {
        "summary": "Пользователи системы (admin)",
        "security": [{ "bearerAuth": [] }],
        "responses": { "200": { "description": "OK" } }
      }
    }
  }
}